Default level for user consent should be Production only (OSIO#2692)
Reorder the `FeatureLevel` constants, with `unknown` becoming the
highest value for the comparisons.
Refactor the `ComputeEnablementLevel` to put the logic of comparing
the internal/external level in it and set the default value when
the feature is disabled or has no strategy configured (or an invalid
value)
Refactor some functions to avoid returning pointers.
Include a test to verify the enablement level for a feature with
a misconfigured level and for a user with no consent.

Fixes openshiftio/openshift.io#2692

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>